Fix schemas caching issue
Caching didn't work properly and was keep adding new entries for the
same schemas.

The underlying problem was with how ruby handles hash key lookups.

https://ruby-doc.org/core-2.2.0/Hash.html#class-Hash-label-Hash+Keys

```
Two objects refer to the same hash key when their hash value is
identical and the two objects are eql? to each other.
```

`SchemaKey` and `SchemaVer` were only implementing `==`, but not `eql?` and
`hash`.

We could fix the problem by adding those methods, but more elegant is
using Struct class: https://ruby-doc.org/core-2.2.0/Struct.html

Struct gives `==`, `eql?` and `hash` so it can be used as a compound hash key.