More resilient schema lookup
It happened to me once or twice that specs were running for a long time,
and then it didn't pass b/c there was a problem with iglucentral
connection.

I don't have the exception at hand, and couldn't reproduce it, but I believe it was a flaky network
connection or maybe iglucentral wasn't available. I've added `read_timeout` and simple retry in case it happens again.

Hopefully this will make it more robust.